# Perf: Fix slow ACL query