@torkelo 2019-06-04
Did you test it with this test data? https://github.com/grafana/grafana/pull/17422 

Did you notice that the GetSystemStats query also gets slow?