@torkelo 2019-06-04
I also noticed the slow login time