@aocenas 2019-06-04
Added unit testes (tested with previous implementation) so it seems like it works the same so far. In any case some more reviewing is probably in order as this is kinda sensitive query that is added to multiple other queries.