@torkelo 2019-06-05
Tested this now and it does fix both the login & seach issue for me and seems to keep the old behavior / logic