@aocenas 2019-06-05
@torkelo regarding the GetSystemStats query I do not see a huge slowdown with lots of dashboards. The queries there do not use the ACL check subquery. There is some dashboard - dashboard_acl joins that could get slower with number of dashboards but probably not significantly and it should be easily cached by the DB.