@torkelo 2019-06-05
Regarding GetSystemStats ,  I was reading the output wrong it took 18ms not 18 seconds